Experimental timestamp return value from commands
In the track package it was discovered that accurate knowledge of the
time when the mount encoders are read is important for accurate
pointing error determination. Normally in the track package this time is
measured after the get_enq() method returns, but multiple lines of
Python code are executed between the time the datagram from Gemini is
received and the return of this method, so timing would be improved if
it is measured as close as possible to when recv() is called. This
commit applies such changes, with corresponding changes in the track
package. However no correlation between the error in the blind error
source and the slight improvement in timing synchronization as applied
here were found, so it is suspected that the impact of these changes
is negligible. However this could be revisited in the future, hence this
experiment is preserved here.