Restrict pySerial version due to API changes
The pySerial project made a change to the parameter name for the
`read_until()` method in version 3.5. Since this project apparently does
not follow semantic versioning (made an API-breaking change without
incrementing the major version number) I have to restrict the versions
of this package that are allowed.

See also https://github.com/pyserial/pyserial/pull/356 where the
possibility of breaking code was even noticed by the reviewer but then
decided to accept the pull request anyway. Why?? :(